+++ src/three-d-secure/shared/constants.js
@@ -1,5 +1,9 @@
 'use strict';
 
 module.exports = {
-  LANDING_FRAME_NAME: 'braintreethreedsecurelanding'
+  LANDING_FRAME_NAME: 'braintreethreedsecurelanding',
+  CARDINAL_SCRIPT_SOURCE: {
+    production: 'https://songbird.cardinalcommerce.com/cardinalcruise/v1/songbird.js',
+    sandbox: 'https://songbirdstag.cardinalcommerce.com/cardinalcruise/v1/songbird.js'
+  }
 };