+++ test/helpers/promise-helper.js
@@ -1,9 +1,20 @@
 'use strict';
 
+var Promise = require('../../src/lib/promise');
+
 function rejectIfResolves() {
   throw new Error('should not resolve');
 }
 
+function wait(time) {
+  time = time || 1;
+
+  return new Promise(function (resolve) {
+    setTimeout(resolve, time);
+  });
+}
+
 module.exports = {
-  rejectIfResolves: rejectIfResolves
+  rejectIfResolves: rejectIfResolves,
+  wait: wait
 };